more sped on unit test.

+++ tests/api/_bootstrap.php
@@ -6,9 +6,14 @@ $app->setRequestForConsoleEnvironment();
 
 $app->boot();
 
-Artisan::call('migrate:reset');
-Artisan::call('migrate');
-Artisan::call('db:seed');
-Artisan::call('es:uninstall');
-Artisan::call('es:install');
+if(!defined('setup')) {
+    Artisan::call('migrate:reset');
+    Artisan::call('migrate');
+    Artisan::call('db:seed');
+    Artisan::call('es:uninstall');
+    Artisan::call('es:install');
+Log::info('##############');
+    define('setup', TRUE);
+}
+
 Mail::pretend(true);
